Remove internal use of string attr in loader option
Fixed issue where a "removed in 2.0" warning were generated internally by
the relationship loader mechanics.

This changeset started the effort of converting all string usage
in the test suite, however this is a much longer job as the
use of strings in loader options is widespread.  In particular
I'm not totally comfortable with strings not being accepted
in obvious spots like Load(User).load_only("x", "y", "z"), which
points to a new string expecting functionality that's not
what's there now.  However at the moment it seems like we need
to continue removing all support for strings and then figure out
"immediate strings from an explicit class" later.

Fixes: #6115
Change-Id: I6b314d135d2bc049fd66500914b772c1fe60b5b3