Commentary; run criteria.params() if statement isn't cached?
Considering adjustment to 56f9c7743e9083add69a10501a503f,
if statement is not cached, skip the relatively expensive
step of re-processing the criteria clause.

However, this causes the overall cache key of the statement
to come out differently which should also be avoided.

Likely we would not merge the actual change, just the
comment here.

References: #6139
Change-Id: Idb555b78d8d7950d084315e004448f64cf59bb5c