Add support for 'code' in deprecation warnings
SADeprecationWarning and SQLAlchemyError now share 'code' and
'_code_str' through a small mixin/helper class.

In the case of the deprecation warnings, the code_str is injected into
`__str__`, which means the message will be built when a warning is
emitted. An alternative approach would be to build a distinct message
when the warning is instantiated.

`s9r1` is a new error code which points directly to the documentation
on backref cascades, and this is now in place as the code for
`cascade_backrefs` warnings.

Fixes: #6148