apply **kw to SchemaEventTarget._set_parent
Fixed issue in the ``SchemaEventTarget._set_parent`` method that was
missing ``**kw`` in its argument signature which was added to this private
method as part of :ticket:`6152`. The method is not invoked internally, but
could potentially be used by a third party system, as well as that the lack
of a proper signature could be misleading in debugging other issues.

Change-Id: I00dce3e533183969b09946f3e918cb30e1154623
References: #6282