Ensure sparse backend selection sorts on driver also
Because we have --dbdriver now, we usually have
lots of db drivers queued up which affect the selection
of backend tests.   the sort in possible_configs_for_cls()
is used so that when we are only picking some backends
due to the sparse flag, we get the same answer each time
otherwise pytest-xdist raises an error if two workers
have different tests.

adding the sort for driver fixes an issue on jenkins
master build where the memusage test run frequently failed
due to mismatched test collection.

Change-Id: I8584c1a3a4e5b759c0e5a2602a93432d9cc8168d