add optional proxy_class to track w/ proxy_key
Fixed regression in ORM where using hybrid property to indicate an
expression from a different entity would confuse the column-labeling logic
in the ORM and attempt to derive the name of the hybrid from that other
class, leading to an attribute error. The owning class of the hybrid
attribute is now tracked along with the name.

Fixes: #6386
Change-Id: Ica9497ea34fef799d6265de44104c1f3f3b30232